Modularize functions
