+++ src/modules/getScores.js
@@ -1,0 +1,6 @@
+const getScores = async (url) => {
+  const response = await fetch(url)
+  return response.json()
+}
+
+export default getScores
