+++ src/modules/submitScores.js
@@ -1,0 +1,15 @@
+const submitScores = async (name, score, url) => {
+  const response = await fetch(url, {
+    method: 'POST',
+    body: JSON.stringify({
+      user: name,
+      score: score
+    }),
+    headers: {
+      'Content-type': 'application/json; charset=UTF-8',
+    },
+  })
+  return response.json()
+}
+
+export default submitScores
